requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
numpy==2.2.5
opencv-python==4.11.0.86
PyYAML==6.0.2
